Update junk code

+++ JunkCode.h
@@ -7,11 +7,11 @@
 		float pkWC = -0.1890433F;
 		short qMFf = -13863;
 		float Scz = 0.0003353103F;
-		long blq = 1894440182021677;
+		long long blq = 1894440182021677;
 		int loQZ = 4600;
 		short SSg = -569;
 		double Nlc = 7.688929E-38;
-		long ITJX = 85715168170936;
+		long long ITJX = 85715168170936;
 		short YJW = 5821;
 		float mLZp = -4.782804E-23F;
 		float Rcte = -3.295265E+35F;
@@ -21,12 +21,12 @@
 		if (fiP == 117537) {
 			fiP = 101141;
 		}float MzVq = -3.575458E-31F;
-		long SqHX = 39279194791865266;
+		long long SqHX = 39279194791865266;
 		byte Flu = 63;
 		int fjdg = 3473;
 		if (fjdg == 132367) {
 			fjdg += 448497;
-		}long zaML = 71508859030974882;
+		}long long zaML = 71508859030974882;
 		byte hjL = 190;
 		byte kyq = 98;
 		float nzoU = -9.699346E-20F;
@@ -42,9 +42,9 @@
 	}void yAD() {
 		float Ukie = 1.006341E+22F;
 		short SHH = -24457;
-		long iOjH = 78123017641534572;
-		long WxNt = 14962172374922929;
-		long PQXp = 49303766082865504;
+		long long iOjH = 78123017641534572;
+		long long WxNt = 14962172374922929;
+		long long PQXp = 49303766082865504;
 		float DOP = -72.65849F;
 		byte fZA = 55;
 		float cPWG = -2.456231E+13F;
@@ -55,7 +55,7 @@
 		byte JKj = 239;
 		float jgs = -1.42428E+13F;
 		byte NBs = 125;
-		long AXb = 55193463031154378;
+		long long AXb = 55193463031154378;
 		short ORT = -6611;
 		byte eeRY = 68;
 		short cqMJ = 17474;
@@ -73,8 +73,8 @@
 		int jIGl = 500534;
 		if (jIGl == 500534) {
 			jIGl = jIGl + 338685;
-		}long TFW = 50878517082621433;
-		long UJRP = 51531202203453629;
+		}long long TFW = 50878517082621433;
+		long long UJRP = 51531202203453629;
 		int scTN = 245470268;
 		byte zXz = 241;
 		short SgL = 3053;
@@ -82,21 +82,21 @@
 		double toFB = 6.048314E+29;
 		double wzE = 4.407349E-32;
 		int hUfF = 417911;
-		long mmd = 15097932672052197;
+		long long mmd = 15097932672052197;
 		short CkdA = -21494;
 		byte qQzH = 201;
-		long qqQ = 3055800610195636;
+		long long qqQ = 3055800610195636;
 		int AVco = 4590;
-		long ZsZo = 34184900361128444;
-		long Ojm = 5536304765212537;
+		long long ZsZo = 34184900361128444;
+		long long Ojm = 5536304765212537;
 		double IpE = -4.019571E-20;
 		double mLN = 9.229347E-13;
 		byte Mcy = 193;
-		long WPln = 63100335683437728;
-		long JlP = 74327779325384321;
+		long long WPln = 63100335683437728;
+		long long JlP = 74327779325384321;
 		double tHs = 2.400811E+36;
 		float ydo = 3.500335E+11F;
-		long kACa = 28701185323580146;
+		long long kACa = 28701185323580146;
 
 	}void LlKk() {
 		double Tmjj = -7.061864E-37;
@@ -112,11 +112,11 @@
 		float ktW = -3.086317E-15F;
 		short qfD = -4552;
 		double xIz = 9.070274E-16;
-		long kjDC = 2133940111596466;
+		long long kjDC = 2133940111596466;
 		double GwG = 2.731126E+17;
 		float hyEe = -2.279751E+11F;
 		double Rpz = 9.552808E+21;
-		long oEo = 82530187182409332;
+		long long oEo = 82530187182409332;
 		int smG = 19229677;
 		if (smG == 267521) {
 			smG = 236509;
@@ -130,7 +130,7 @@
 		int REFA = 316778866;
 		double KiT = -3.746298E-12;
 		byte Pys = 186;
-		long hyhg = 57456949058999326;
+		long long hyhg = 57456949058999326;
 		short yhk = 17442;
 		float NChU = 1142.811F;
 		double CmCO = -1.052683E+36;
@@ -139,16 +139,16 @@
 	}void AfUh() {
 		double WRm = -1.571605E+19;
 		float Jue = 2.406502E+22F;
-		long BKnl = 28839808079347339;
+		long long BKnl = 28839808079347339;
 		byte GOfD = 95;
 		int Ksu = 482545;
 		double Aps = -6.417854E+10;
 		double tHw = -1.505777E+32;
 		int ZQhX = 85384822;
-		long wLz = 30096572094597197;
-		long pMl = 43588899894936715;
+		long long wLz = 30096572094597197;
+		long long pMl = 43588899894936715;
 		float pei = 7.593322E+28F;
-		long iOlj = 56753801881732966;
+		long long iOlj = 56753801881732966;
 		double dIZt = -1.423682E-30;
 		float aQbW = 7.548684E-13F;
 		byte RygU = 178;
@@ -178,7 +178,7 @@
 		if (muDZ == 8530) {
 			muDZ = muDZ + 137581;
 		}short GMLf = 25315;
-		long InX = 85429810336775711;
+		long long InX = 85429810336775711;
 		short yaz = 21682;
 		short Oop = 31722;
 		int RLe = 26559393;
@@ -189,7 +189,7 @@
 			AFnT = AFnT + 289744;
 		}byte gihj = 5;
 		float JFc = -5.770137E+14F;
-		long xQUA = 66257866024165558;
+		long long xQUA = 66257866024165558;
 		float PUOw = 2.100696E+24F;
 		double tdd = -3.306713E+16;
 		float uDm = 3.889275E-31F;
@@ -203,7 +203,7 @@
 		byte TXe = 129;
 		float NPMx = -9.530908E-27F;
 		short foUV = 3417;
-		long xIp = 30104015784956945;
+		long long xIp = 30104015784956945;
 		short SgkV = 10082;
 
 	}void BNxW() {
@@ -219,22 +219,22 @@
 		}float aXk = 184201;
 		if (aXk == 184201) {
 			aXk = 422778;
-		}long FEf = 62685377982856870;
+		}long long FEf = 62685377982856870;
 		float aiiK = -1.084086E-16F;
 		byte BHFV = 40;
 		double FbfN = -1.951803E-37;
 		short FgpW = -22296;
-		long nMle = 15422234801584102;
+		long long nMle = 15422234801584102;
 		short wyKj = -31412;
 		int DfEx = 168734;
 		if (DfEx == 103922) {
 			DfEx += 866810;
-		}long cKL = 33736566959861461;
+		}long long cKL = 33736566959861461;
 		float PXV = 1.113123E+21F;
 		double ScD = 1.524247E+28;
 		float HJR = -2.451151E+08F;
 		int eKa = 84;
-		long aOyR = 44058878500468963;
+		long long aOyR = 44058878500468963;
 		double ykVF = 7.34182E-25;
 
 	}void Wchh() {
@@ -248,11 +248,11 @@
 		byte IhR = 80;
 		byte gdQ = 172;
 		double qQic = 7.326244E-34;
-		long Ksjo = 50006926400305903;
+		long long Ksjo = 50006926400305903;
 		int YDD = 85;
-		long QqB = 50401719361946772;
+		long long QqB = 50401719361946772;
 		int EzZZ = 24;
-		long FDqW = 37040454287466507;
+		long long FDqW = 37040454287466507;
 		int xQT = 863133;
 		if (xQT == 722369) {
 			xQT = xQT + 813579;
@@ -265,7 +265,7 @@
 		short oafs = -16357;
 		float IHpp = -1.878297E-11F;
 		float zqi = 7.99431E-16F;
-		long wnU = 17113993331615476;
+		long long wnU = 17113993331615476;
 		float IUO = -7.314778E+35F;
 		float UZQX = 0.283686F;
 
@@ -278,12 +278,12 @@
 		int DCUy = 83;
 		if (DCUy == 719605) {
 			DCUy += 502420;
-		}long qTF = 16332674576460631;
+		}long long qTF = 16332674576460631;
 		int OQVk = 910546255;
 		if (OQVk == 949298) {
 			OQVk += 237041;
 		}float jLXC = 2.244448E+14F;
-		long CmmJ = 6862231715344241;
+		long long CmmJ = 6862231715344241;
 		int pZh = 26;
 		if (pZh == 709614) {
 			pZh = 228531;
@@ -300,7 +300,7 @@
 		short hUZ = -15241;
 		short EYKM = 22218;
 		float IHHV = 3.536026E-25F;
-		long xyju = 44834559216076073;
+		long long xyju = 44834559216076073;
 		short ClHV = 4364;
 
 	}void DbL() {
@@ -310,7 +310,7 @@
 		short czFE = 16385;
 		short qHI = -12172;
 		short eWJP = -10161;
-		long AEMz = 24261592836869879;
+		long long AEMz = 24261592836869879;
 		double wBs = -3.808327E+23;
 		double uNMt = 5.444823E+35;
 		double hFfO = 4.439228E-39;
@@ -320,11 +320,11 @@
 			yBK = 337899;
 		}float gdJ = 2.647583E-08F;
 		double TofQ = -1.327815E+21;
-		long mzMg = 15607181456056947;
+		long long mzMg = 15607181456056947;
 		byte VWW = 89;
 		byte Ohmo = 140;
 		byte kCT = 68;
-		long LhJQ = 76050607600641673;
+		long long LhJQ = 76050607600641673;
 		short WAi = 4785;
 		int FNpq = 9288;
 		if (FNpq == 686387) {
@@ -336,17 +336,17 @@
 		byte WlO = 52;
 		double qZY = 721136.8;
 		float KjnL = 0.0004606611F;
-		long lNu = 1736199152432162;
+		long long lNu = 1736199152432162;
 		int zxU = 1274;
 
 	}void dmfc() {
 		float Agld = 1.236583E-07F;
-		long UVo = 30660076281559074;
+		long long UVo = 30660076281559074;
 		double yOqM = 3.20273E-27;
-		long DuHu = 62185531227518095;
+		long long DuHu = 62185531227518095;
 		byte ikc = 157;
 		short Kkbo = 459;
-		long Ebeq = 9472929235300952;
+		long long Ebeq = 9472929235300952;
 		int iUNT = 881860;
 		if (iUNT == 810989) {
 			iUNT += 691503;
@@ -355,7 +355,7 @@
 			tpPe = tpPe + 32891;
 		}byte Htk = 149;
 		int yjol = 6655;
-		long bsj = 32943505234645612;
+		long long bsj = 32943505234645612;
 		int Lpda = 2865;
 		unsigned TwAt = 638926;
 		if (TwAt == 638926) {
@@ -368,10 +368,10 @@
 		byte hhlJ = 155;
 		int jaWJ = 9420;
 		short pToF = -26391;
-		long nUp = 83414690227742133;
+		long long nUp = 83414690227742133;
 		int TDft = 389989898;
 		byte IOBW = 107;
-		long txtM = 15756375987883409;
+		long long txtM = 15756375987883409;
 		double BHX = 3.360405E+19;
 
 	}void tXm() {
@@ -383,7 +383,7 @@
 		if (fQq == 212355) {
 			fQq += 496660;
 		}short zej = -9915;
-		long enl = 1060735038088146;
+		long long enl = 1060735038088146;
 		short xcW = 31146;
 		double spgM = -2.579012E+08;
 		float oFqY = 8.793466E-38F;
@@ -391,20 +391,20 @@
 		double hzh = 11037.38;
 		short bqom = 9397;
 		short JxYj = 16943;
-		long xxMf = 55121165740327137;
+		long long xxMf = 55121165740327137;
 		float SsH = 1.097008E+13F;
 		double TjJo = -1.121369E+22;
-		long aGMA = 54448672419249767;
+		long long aGMA = 54448672419249767;
 		double aSH = -6.091222E-21;
 		float SBW = -1.825021E-09F;
 		double tLzW = 1.478602E-30;
-		long eSsa = 48359426684988857;
+		long long eSsa = 48359426684988857;
 		byte whyj = 137;
 		float tIjc = -2.073869E+21F;
 		int VYw = 51;
 		if (VYw == 671153) {
 			VYw = VYw + 634193;
-		}long FaHZ = 18998823982732167;
+		}long long FaHZ = 18998823982732167;
 		byte zQzT = 216;
 		byte VLC = 172;
 
@@ -413,14 +413,14 @@
 		int COgI = 6539;
 		float ymG = 1.421139E-20F;
 		short XgsQ = 14204;
-		long Hoc = 62227077685140300;
+		long long Hoc = 62227077685140300;
 		int bHJh = 861;
 		if (bHJh == 292932) {
 			bHJh = bHJh + 658477;
 		}short xsOa = 19803;
 		double HHZ = -5.55913E+23;
 		float Beff = 5.937965E-23F;
-		long apz = 75431418638157057;
+		long long apz = 75431418638157057;
 		float AxI = -5.023227E-37F;
 		byte azT = 69;
 		short VZSW = -26430;
@@ -438,24 +438,24 @@
 		int nsJ = 976053948;
 		if (nsJ == 913734) {
 			nsJ = 475585;
-		}long ICY = 56709098344882603;
-		long OSes = 18071926217469932;
+		}long long ICY = 56709098344882603;
+		long long OSes = 18071926217469932;
 		float VLz = -1.124415E-14F;
-		long mmze = 64259874908381494;
+		long long mmze = 64259874908381494;
 		short tSl = -17875;
 		float tJu = -0.0007293841F;
 
 	}void qmY() {
-		long YpA = 21153766403761411;
+		long long YpA = 21153766403761411;
 		short amu = -8356;
 		byte Waj = 58;
 		double yeu = 3.157138E+15;
 		double bwhN = -2.539548E+35;
-		long UeVZ = 26056167748345628;
+		long long UeVZ = 26056167748345628;
 		float dfg = 2.703573E+13F;
 		double MVCO = -3.751478E-25;
 		int aEbK = 9804;
-		long ESA = 24176580170886608;
+		long long ESA = 24176580170886608;
 		double ayef = -4.207703E+33;
 		double DQz = -6.303266E-10;
 		double UOK = 4.383691E+23;
@@ -466,7 +466,7 @@
 		double fup = -9.505898E+31;
 		double zLf = -2.444882E-26;
 		byte OfY = 115;
-		long EXl = 29158939643922049;
+		long long EXl = 29158939643922049;
 		int bBN = 85;
 		if (bBN == 471385) {
 			bBN = 180331;
@@ -485,13 +485,13 @@
 
 	}void MYa() {
 		double mLgH = -2.089218E+19;
-		long cQGt = 86173131169455417;
+		long long cQGt = 86173131169455417;
 		int mtby = 70439195;
 		if (mtby == 391324) {
 			mtby = mtby + 855899;
 		}float qwKC = 2987707;
 		short VCnb = -17548;
-		long jjup = 16631097800971701;
+		long long jjup = 16631097800971701;
 		double Izn = -0.0155813;
 		byte UyYu = 243;
 		byte ObV = 160;
@@ -517,7 +517,7 @@
 			gdZN += 793489;
 		}short zTV = -24652;
 		float ksNS = -6.130833E-08F;
-		long SSG = 58300843647707842;
+		long long SSG = 58300843647707842;
 		double hJl = -0.004439062;
 		byte TPs = 61;
 		float Nmn = -617.1819F;
@@ -531,24 +531,24 @@
 		int ucqK = 86;
 		if (ucqK == 609956) {
 			ucqK += 577940;
-		}long exOU = 42438153265274121;
+		}long long exOU = 42438153265274121;
 		short Stni = 26918;
 		float EynG = 3.827426E+11F;
 		float VWnd = 1.470801E+08F;
 		short CYVl = 2982;
 		short iFYt = 208;
 		byte sqZ = 185;
-		long ZKW = 69980083078930542;
-		long KgG = 82241253613252526;
+		long long ZKW = 69980083078930542;
+		long long KgG = 82241253613252526;
 		short Ewq = 921;
-		long eHB = 79279788223979561;
+		long long eHB = 79279788223979561;
 		short YjmH = -26429;
-		long uXtP = 15556658642916837;
+		long long uXtP = 15556658642916837;
 		float BzTN = 8.143449E-33F;
 		float yfWu = -2.007504E+07F;
 		double uhl = 2.559295E+27;
-		long CQO = 49501427197550993;
-		long lYV = 14736705012884414;
+		long long CQO = 49501427197550993;
+		long long lYV = 14736705012884414;
 		short CnK = 14000;
 		byte AZCo = 194;
 		double hZbc = -2.355351E-09;
@@ -556,7 +556,7 @@
 		double SEih = -1.546682E-33;
 		float RsNk = -8.793682E-31F;
 		float JeL = -3.366191E+17F;
-		long WzP = 10587359023282783;
+		long long WzP = 10587359023282783;
 		int DzB = 35841960;
 		if (DzB == 96483) {
 			DzB += 411238;
@@ -570,19 +570,19 @@
 		double gagU = 3.364965E+14;
 		short bApm = 22853;
 		short EINN = 28096;
-		long zth = 85022728680694227;
+		long long zth = 85022728680694227;
 		int iQs = 76795264;
 		double JESV = 2.909744E-30;
-		long JdL = 74811650820992883;
+		long long JdL = 74811650820992883;
 		int DYN = 3680;
 		double ejqw = -0.0002097897;
 		short FlZM = -17536;
 		double YODb = -1.415881E-12;
-		long DOK = 22245027440013715;
+		long long DOK = 22245027440013715;
 		byte bkgM = 126;
-		long PdZ = 60917792228461663;
-		long ezb = 72116508129161993;
-		long zBw = 67968777366238083;
+		long long PdZ = 60917792228461663;
+		long long ezb = 72116508129161993;
+		long long zBw = 67968777366238083;
 		int AqXP = 14;
 		if (AqXP == 344365) {
 			AqXP += 255401;
@@ -601,12 +601,12 @@
 		int OpMt = 7577;
 		int LEeR = 4509;
 		float PXMZ = -5.862267E-21F;
-		long GuH = 18628923974149592;
+		long long GuH = 18628923974149592;
 		double LEbg = -1.691986E+10;
-		long gPuR = 6760609828647799;
+		long long gPuR = 6760609828647799;
 		short ObpG = 29290;
 		float xPo = 2.638826E-19F;
-		long wlE = 48355138350733961;
+		long long wlE = 48355138350733961;
 		byte pnwo = 254;
 		int KafH = 94;
 		short wFGE = 27142;
@@ -618,7 +618,7 @@
 		if (IJU == 620106) {
 			IJU = IJU + 813420;
 		}short fVA = 28583;
-		long oXo = 5949403130211973;
+		long long oXo = 5949403130211973;
 		float mDfj = -7.044246E-34F;
 		byte jxS = 29;
 		float hEI = -3.386724E+20F;
@@ -632,34 +632,34 @@
 		double ocCE = 3129.956;
 		byte BScU = 120;
 		float fAw = -4.679588E+18F;
-		long tbf = 29698977505506909;
+		long long tbf = 29698977505506909;
 		double niMQ = -7.821971E-15;
-		long jiY = 11518390061774619;
+		long long jiY = 11518390061774619;
 		float BQNc = 1566.071F;
-		long JUw = 23606569345534718;
-		long fSY = 23961765888401567;
+		long long JUw = 23606569345534718;
+		long long fSY = 23961765888401567;
 		double LIEL = 7.15565E-07;
 		double cFCT = -1.251711E-26;
-		long QXGH = 70223582957779391;
+		long long QXGH = 70223582957779391;
 		double BBmz = -4.089496E-36;
 		double egVI = -6.41404E+09;
 		double eZM = -2.428021E+10;
 		double PUs = 4073.067;
 		float BWj = -2.349583E+26F;
-		long NdKH = 28856299094263230;
+		long long NdKH = 28856299094263230;
 		float CyZy = -5.695562E+09F;
 		byte OLXU = 111;
 		float oOE = 8.386542E-23F;
-		long KcUX = 32880515848342789;
+		long long KcUX = 32880515848342789;
 
 	}void SieU() {
-		long WgW = 14376797122487021;
+		long long WgW = 14376797122487021;
 		int zmo = 459798;
 		if (zmo == 947921) {
 			zmo = 212168;
 		}byte ukQ = 99;
 		float iLSV = 1.63155E-30F;
-		long dxz = 88987203666825721;
+		long long dxz = 88987203666825721;
 		double HmED = -1.697107E+19;
 		int mUGE = 173584464;
 		byte Ktnc = 8;
@@ -673,17 +673,17 @@
 			YSW += 294651;
 		}float QRR = 1.809516E-18F;
 		byte jtB = 186;
-		long oBGQ = 50745450251934515;
-		long RfEY = 21723371251363991;
+		long long oBGQ = 50745450251934515;
+		long long RfEY = 21723371251363991;
 		short mNcL = 8513;
 		short Xew = 21010;
 		short CNu = 26217;
-		long lJfc = 12872883580260708;
+		long long lJfc = 12872883580260708;
 		double hsb = 1.323309E-34;
 		byte kpE = 233;
 		double aWw = 1014.562;
 		float QpTy = 3.383268E+13F;
 		short Ope = 9519;
-		long hkdA = 83783881562154232;
+		long long hkdA = 83783881562154232;
 
 	}
